altos-avr: Store TM tick and state in log, get logging on/off working
To sync TM and TS data after flight, store the TM tick data in the log
along with the TM state. This should provide sufficient data to
synchronize the two data streams.

Use the TM state to enable/disable logging, log from boost to landing
and otherwise do not log.

Signed-off-by: Keith Packard <keithp@keithp.com>