altos: Delay reboot by a second to avoid re-entering idle mode
TM and Tm go into idle mode if they receive a packet after boot
time. When tebooting the device over the packet link, the packet
master would be (rapidly) polling the device for additional data and
so the device would invariably receive a packet during bootup and go
into idle mode again.

Delay the reboot by a second to give the controller time to disable
the master end of the packet link.

Signed-off-by: Keith Packard <keithp@keithp.com>