altos/cc1111: Adjust receive parameters to improve sensitivity
This removes the packet quality test and carrier sense tests when
deciding whether to start decoding a packet. This lets more bad
packets through, but the CRC check catches those and now we're
regularly receiving packets down to -110 or even -112 dBm. Before this
change, we'd rarely see packets as low as -105dBm

Signed-off-by: Keith Packard <keithp@keithp.com>