altosui/telegps: Undo the frequency/telemetry menu changes
Using menus for these items isn't what we want; we want to show the
current value on the screen, which is what a combo box is for. Switch
back to those and remove the frequency and telemetry config from the
TeleGPS menu bar so that Mac OS X is happy.

Signed-off-by: Keith Packard <keithp@keithp.com>