micropeak/altosui/telegps: Fix icon file names
The icon file names are now structured to have an organization name
and application name component. This should avoid collisions with
other packages when the icon files are installed in a shared
directory, as on Linux. Within the Java .jar file, the new names need
to be correctly referenced or we won't find them.

Signed-off-by: Keith Packard <keithp@keithp.com>