altoslib: Stop re-loading previous flight when starting telemetry monitoring
This feature is necessary on AltosDroid because there's no other UI
for reloading an old flight. It's not useful on the desktop where you
really just want to see new telemetry; old telemetry is availbale
through the 'graph flight' or 'replay' mechanisms.

AltosDroid uses a different class for reading telemetry data, so we
can just remove this feature from altoslib at this point.

Signed-off-by: Keith Packard <keithp@keithp.com>