altos: Make sure we drop the SPI mutex when aborting cc1120 recv
The cc1120 receive code is 'twisty', in that it acquires the SPI mutex
in the middle of an interrupt handler so that it can quickly start the
radio when the signal comes in. This means we have to be careful to
drop that mutex in all paths which leave the receive code, including
when the packet receive is aborted mid-way through the process.

Signed-off-by: Keith Packard <keithp@keithp.com>