altosui: Remove the dregs of AltosDroid load-old-telem code
AltosDroid used to scan the old .telem file to return to the previous
flight state on restart. Now it just loads the old state object
instead, a vast improvement in performance.

To do that, there were some changes in the altoslib code
required. This patch just removes those, fixing replay bugs in TeleGPS
along the way.

Signed-off-by: Keith Packard <keithp@keithp.com>