altoslib: Mark listener as 'not running' on EOF.
This adds a 'running' member to the AltosListenerState class, and when
the replay reader reaches EOF, marks the listener as no longer
running.

AltosUI and TeleGPS now display 'done' in the 'Age' field when this
occurs, to let the user know that the replay is over.

Also make sure that the display timers are stopped when this happens,
or when the window is closed.

Signed-off-by: Keith Packard <keithp@keithp.com>