altosdroid: Lots of bluetooth connection changes
Appears to more reliably abort in-progress connection attempts so you
can switch TBT devices without having the previous device in
operation.

Shows which device the connection is being attempted for.

Eliminate the 10-second timer and just disable the service when the GUI
shuts down while no BT connection is running.

Signed-off-by: Keith Packard <keithp@keithp.com>