altos/telelco: Add drag race UI
With the unit disarmed, press and hold the fire button for five
seconds to enable drag race mode.

The display will show 'dr' for five seconds and beep five times to
indicate that drag race mode is enabled. The decimal points in the
display will all be displayed as an additional visual aid. Once every
five seconds, it will beep.

With drag race mode enabled, you can select a box/pad pair and press
the 'fire' button to add it to the drag race group. For the current
box, all members of the drag race group will have their continuity
LEDs blink slowly. There will be no indication of continuity in this
mode; you'll want to check that before enabling drag race mode. If you
want to de-select a member of the group, just press the fire button
again. Each time you push the fire button, it will beep out the pad
number added or removed.

Arm the box and you will not be able to add or remove members from the
drag race group. Firing will simultaneously fire all members of the
drag race group.

To disable drag race mode, press and hold the fire button for two
seconds. It will beep twice and turn off the decimal points in the display.

Signed-off-by: Keith Packard <keithp@keithp.com>