altosui/telegps/micropeak: Handle both MULTI_LIB and non-MULTI_LIB builds
On x86 systems with -m32 and -m64 compilers, both libaltos32.so and
libaltos64.so are built. Otherwise, we should use libaltos.so

altosui only dealt with libaltos32.so and libaltos64.so, so it only
worked on MULTI_LIB systems. telegps and micropeak only used
libaltos.so, so they wouldn't work correctly on multi-lib systems.

Fix all of them to work either way.

Signed-off-by: Keith Packard <keithp@keithp.com>