altos: cc1200: use FIFO threshold pin output for APRS buffering
Instead of polling the device for fifo space, just use the available
pin configuration to figure out if there is enough space for a single
APRS buffer. Then set the APRS buffer size to match the fifo threshold
setting in the chip so that we know we can write the whole APRS buffer
once the pin says there's space.

Signed-off-by: Keith Packard <keithp@keithp.com>