altoslib: Allow gps time later than requested if it's first
When generating a KML file, we want to position markers near the start
of the flight section. This is done by looking for a GPS coordinate
'before' the starting point of the flight, which doesn't work well
when the first GPS coordinate is later than that. Pick the first point
after the chosen time if there isn't an earlier one.

Signed-off-by: Keith Packard <keithp@keithp.com>