altoslib: Save separate config for local and remote. Use in idle
When using the remote link, there are two separate configuration data
blocks, that for the local device and for remote. Make the link report
both versions, depending on whether it is in remote mode or not.

Request config data in remote mode when running idle monitoring so
that the presented data is for the remote device, not the local one.

Signed-off-by: Keith Packard <keithp@keithp.com>