altos: Share common logging code. Deal with corrupt initial flight records
Move common logging APIs from per-format files into ao_log.c. Then,
change that code to check the first log record in a slot (containing
the flight number) to see if it's invalid and deal with it. That
involves not re-using that slot, and allowing it to be erased.

Corrupted log blocks are reported with a negative flight number.

Signed-off-by: Keith Packard <keithp@keithp.com>