altos/scheme: Work around gcc 7.2.0 optimization bug in memory manager
After marking a set of memory chunks, it's possible that all of them
will be packed tight against 'top', in which case none of them will be
moving. In that case, gcc 7.2.0 appears to generate incorrect code
causing the loop to be abandoned, meaning that we don't actually
collect anything at all.

Add a quick short-circuit test just after the mark phase that skips
the code which wouldn't do anything in this case.

Signed-off-by: Keith Packard <keithp@keithp.com>