altos/scheme: apply also needs to not free value list on lambdas
When apply is invoked on any function, the cons in the argument list
cannot be immediately freed as they have been passed to the
function. That applies to both built-ins as well as lambdas; this
patch removes the special ao_scheme_skip_cons_free global and just
marks the stack in both cases.

Signed-off-by: Keith Packard <keithp@keithp.com>