libaltos: Use case-insensitive compare when matching BT MACs
We use the BT MAC vendor portion to figure out which port to connect
to as that is simpler and takes less network traffic than actually
doing discovery. However, on Windows, we were generating the address
in lower case and comparing against upper case vendors, which didn't
work out too well.

Signed-off-by: Keith Packard <keithp@keithp.com>