micropeak: Make reported 'max height' value show filtered max height
MicroPeak tries to filter out bogus pressure data so that the value
reported on the LED is reasonable. We want to report that same value
in the UI, which means replacing the normal mechanism for computing
max height from the time series of height data with the value computed
from the saved minimum pressure value recorded in the device.

Signed-off-by: Keith Packard <keithp@keithp.com>