altosui: Fetch RF calibration value for TBT v4.0 units from web
We shipped a pile of TBT v4.0 units without programming the RF
calibration value into flash. This change checks to see if the target
is in the right serial number range and has the default RF calibration
value, in which case it fetches the RF calibration value from our web
site.

Signed-off-by: Keith Packard <keithp@keithp.com>