doc: Describe Packet Link usage and selecting RF param defaults for TeleMini
Packet Link mode is the only communication method available for
TeleMini and deserves some special mention in the docs.

TeleMini allows forcing of RF parameters back to known defaults by
hooking SWCLK to ground at power on (hole 7 to hole 3 on the debug
connector). Describe this process so that users will be able to get
TeleMini back to known values and recover.

Signed-off-by: Keith Packard <keithp@keithp.com>