altos: Fix ISR declarations to make them non-weak
In my zeal to eliminate warnings, I added declarations of all ISR
functions to a shared header file. However, I managed to include the
'weak' declaration, which meant that the intended ISR functions were
as weak as the 'default' ISR functions. This left all interrupts
non-functional, which doesn't make for a happy program.

Signed-off-by: Keith Packard <keithp@keithp.com>