altoslib: Linker script changed -> the USB desc is two bytes off
The USB descriptor used to be at 0x0800110c in previous releases and
is now at 0x0800110a, presumably because the linker script changed the
padding requirements of those sections.

Search forward and backwards two bytes to see if we can't find the
descriptors by checking for the descriptor value and size at each
location.

Signed-off-by: Keith Packard <keithp@keithp.com>