altos: Make debug telemini firmware a bit easier to build
To enable debugging, we need to stop using the SWD GPIO pins, which
are normally used to trigger boot-time firmware behavior -- SWDIO is
used to force entry to the boot loader while SWDCLK is used to force
default frequency and callsign.

This patch makes it easy to edit ao_pins.h to disable the default
frequency option.

Signed-off-by: Keith Packard <keithp@keithp.com>