altos: add APRS 'offset' value to set APRS transmision time
This value is the offset from the top of the minute for the first APRS
packet in the minute. Subsequent packets will be transmitted
'interval' seconds apart for the rest of the minute.

This allows multiple transmitters to be configured to share the same
frequency and not transmit at the same time.

Note that this offset only works when the device has GPS lock.

Signed-off-by: Keith Packard <keithp@keithp.com>