altos: Round selected log size down to ao_storage_block multiple
There was some (broken) code that complained if the value set wasn't
correct, that has been replaced by code that accepts any value and
just rounds it down to a multiple of ao_storage_block. The code also
stops complaining when set to the current value when storage isn't
empty.

Signed-off-by: Keith Packard <keithp@keithp.com>