altos/stmf0, stm32f0: Call ao_boot_chain to start application
This used to be done inside ao_boot_check_chain or ao_boot_check_pin,
but it made more sense to have it done in _start instead. stm and lpc
were adjusted, but stmf0 and stm32f4 were missed.

This caused boards using this process to wedge in the boot loader.

Signed-off-by: Keith Packard <keithp@keithp.com>