altos/telegps-*: Fix log end discovery at startup
We need to find the first unwritten log block to start appending data,
but the code was actually looking for the first empty 256-byte chunk,
which meant that we'd leave a gap of erased data after the previous
log. AltosUI would stop at that point and not recover the remaining
stored data.

Signed-off-by: Keith Packard <keithp@keithp.com>