altos: Add timeout in MS5607 get_sample code
MISO is supposed to rise when the sample value is ready, but sometimes
we miss the interrupt which should be generated. I spent a day
attempting to fix this, but was unable to make it 100%. Instead, add a
10ms timeout, which is longer than the sample time (8.2ms), and then
read MISO directly. If the interrupt fires, we'll read a bit sooner.

Signed-off-by: Keith Packard <keithp@keithp.com>