altos: Use pyro voltage to check igniters
Instead of using an absolute value for the igniter continuity good
voltage, use 15/16 of the pyro voltage instead. This ensures we scale
with the pyro voltage to find marginal igniters even with high voltage
pyro systems.

This doesn't work on TeleMetrum where we have no pyro voltage
value. On these boards, continue to use a fixed 3.5V value.

Signed-off-by: Keith Packard <keithp@keithp.com>