altos: Add Mosaic GPS support to TeleMega v3+
The Septentrio Mosaic-X5 GNSS module provides exceptional performance
along with reporting data above limits within the uBlox chips.

Support for them as an optional GPS unit, connected via the companion
connector on cs_companion1 (rx) and cs_companion2 (tx). When selected,
the internal GPS unit is not used.

When the external GPS is in use, the regular companion interface is
not available.

Signed-off-by: Keith Packard <keithp@keithp.com>